+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Storage Monitors don't render any quantity text when no item is specified to monitor anymore (raoulvdberge)
 - Fixed bug where disks in Disk Drive didn't respect access type or void excess stacks option (raoulvdberge)
 - Fixed crash in Disk Manipulator (raoulvdberge)
+- Fixed oredict not working (raoulvdberge)
 - You can now shift click Grid Filters into a Grid instead of manually inserting them (raoulvdberge)
 - The Solderer inventory isn't sided anymore (raoulvdberge)